+++ packages/translations/.env.example
@@ -1,0 +1,1 @@
+OPENAI_KEY=sk-
