+++ packages/translations/scripts/translateNewKeys/translateText.ts
@@ -1,0 +1,35 @@
+import dotenv from 'dotenv'
+
+dotenv.config({ path: '../../' })
+
+export async function translateText(text: string, targetLang: string) {
+  const response = await fetch('https://api.openai.com/v1/chat/completions', {
+    body: JSON.stringify({
+      max_tokens: 150,
+      messages: [
+        {
+          content: `Only respond with the translation of the text you receive. The original language is English and the translation language is ${targetLang}. Only respond with the translation - do not say anything else. If you cannot translate the text, respond with "[SKIPPED]"`,
+          role: 'system',
+        },
+        {
+          content: text,
+          role: 'user',
+        },
+      ],
+      model: 'gpt-4',
+    }),
+    headers: {
+      Authorization: `Bearer ${process.env.OPENAI_KEY}`,
+      'Content-Type': 'application/json',
+    },
+    method: 'POST',
+  })
+  try {
+    const data = await response.json()
+    console.log('  Old text:', text, 'New text:', data.choices[0].message.content.trim())
+    return data.choices[0].message.content.trim()
+  } catch (e) {
+    console.error('Error translating:', text, 'to', targetLang, 'response', response, '. Error:', e)
+    throw e
+  }
+}
